requirements.txt
# Deep Learning & Scientific Computing
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
pandas>=1.3.0
scikit-learn>=0.24.2

# Image Processing
Pillow>=8.2.0

# Web Scraping
selenium>=4.0.0
beautifulsoup4>=4.9.3
requests>=2.25.1

# Data Visualization
matplotlib>=3.4.2
seaborn>=0.11.1

# Excel Support
openpyxl>=3.1.0
xlrd>=2.0.1

# Utilities
tqdm>=4.61.0
python-dotenv>=0.19.0

# Optional Dependencies
# Uncomment if needed
# tensorboard>=2.6.0  # for logging experiments
# pytest>=6.2.5      # for testing
# black>=21.6b0      # for code formatting
# flake8>=3.9.2      # for code linting

# System Requirements
python>=3.8.0